Move neovim plugin configurations to their own folder
This allows us to use the more explicit require 'config.$PLUGIN'
instead of require '$PLUGIN', which is ambiguous as it could refer to
both the plugin modules as well as our own configuration files.

Examples which have been fixed include the inconsistent 'trouble-config' and
'telescope-config', which were originally named as such due to namespace
configs. Now they are 'config.trouble' and 'config.telescope', which is
consistent with other configuration files such as 'config.lsp'.